Completed apec_table method to create an emission table. Removed old emission table (created with IDL) with a new one, with the same parameters but created with the new procedure, now called reference_emission_table.fits. Re-created many reference files for test. Added test for the new procedure.
In addition, corrected an error in the input files, apec_fakeit_for_test.pha and bapec_fakeit_for_test.pha, for the testing of the fit procedure as they were created without thermal broadening. This might be the reason why sometimes the test for the fit failed in some systems and may therefore have solved Issue #15.